chore(Matomo): move sdk loader to different file
