+++ src/integrations/Matomo/browser.js
@@ -12,7 +12,7 @@ import {
   checkCustomDimensions,
 } from './util';
 import { getHashFromArrayWithDuplicate } from '../../utils/commonUtils';
-import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
+import { loader } from './loader';
 
 class Matomo {
   constructor(config, analytics, destinationInfo) {
@@ -55,19 +55,7 @@ class Matomo {
 
   loadScript() {
     window._paq = window._paq || [];
-    (function (serverUrl, siteId) {
-      let u = serverUrl;
-      window._paq.push(['setTrackerUrl', `${u}matomo.php`]);
-      window._paq.push(['setSiteId', siteId]);
-      const d = document;
-      const g = d.createElement('script');
-      const s = d.getElementsByTagName('script')[0];
-      g.async = true;
-      u = u.replace('https://', '');
-      g.src = `//cdn.matomo.cloud/${u}matomo.js`;
-      g.setAttribute('data-loader', LOAD_ORIGIN);
-      s.parentNode.insertBefore(g, s);
-    })(this.serverUrl, this.siteId);
+    loader(this.serverUrl, this.siteId);
   }
 
   init() {
